// apps/web-svelte/src/lib/api/generated/model/jobActivityBucket.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { JobType } from './jobType';

export interface JobActivityBucket {
  type: JobType;
  start: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  total: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  failed: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  running: number | string;
}